+++ app/Http/Controllers/ServiceRequestController.php
@@ -9,6 +9,8 @@ use App\Http\Requests\ServiceRequest\UpdateRequest;
 use App\Http\Resources\ServiceRequestResource;
 use App\Http\Resources\ServiceRequestResourceCollection;
 use App\Repositories\Contracts\ServiceRequestRepository;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\Response;
 
 class ServiceRequestController extends Controller
 {
@@ -31,9 +33,12 @@ class ServiceRequestController extends Controller
      *
      * @param IndexRequest $request
      * @return ServiceRequestResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [ServiceRequest::class, $request->get('propertyId'), $request->get('unitId', null)]);
+
         $serviceRequests = $this->serviceRequestRepository->findBy($request->all());
 
         return new ServiceRequestResourceCollection($serviceRequests);
@@ -44,9 +49,12 @@ class ServiceRequestController extends Controller
      *
      * @param StoreRequest  $request
      * @return ServiceRequestResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [ServiceRequest::class, $request->get('propertyId'), $request->get('unitId', null)]);
+
         $serviceRequest = $this->serviceRequestRepository->save($request->all());
 
         return new ServiceRequestResource($serviceRequest);
@@ -57,9 +65,12 @@ class ServiceRequestController extends Controller
      *
      * @param ServiceRequest $serviceRequest
      * @return ServiceRequestResource
+     * @throws AuthorizationException
      */
     public function show(ServiceRequest $serviceRequest)
     {
+        $this->authorize('show', $serviceRequest);
+
         return new ServiceRequestResource($serviceRequest);
     }
 
@@ -69,9 +80,12 @@ class ServiceRequestController extends Controller
      * @param UpdateRequest $request
      * @param ServiceRequest $serviceRequest
      * @return ServiceRequestResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ServiceRequest $serviceRequest)
     {
+        $this->authorize('update', $serviceRequest);
+
         $serviceRequest = $this->serviceRequestRepository->update($serviceRequest,$request->all());
 
         return new ServiceRequestResource($serviceRequest);
@@ -80,11 +94,14 @@ class ServiceRequestController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param ServiceRequest $serviceRequest
+     * @return Response
+     * @throws AuthorizationException
      */
     public function destroy(ServiceRequest $serviceRequest)
     {
+        $this->authorize('destroy', $serviceRequest);
+
         $this->serviceRequestRepository->delete($serviceRequest);
 
         return response()->json(null,204);
