+++ app/Http/Controllers/ServiceRequestOfficeDetailController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\ServiceRequestOfficeDetail\UpdateRequest;
 use App\Http\Resources\ServiceRequestOfficeDetailResource;
 use App\Http\Resources\ServiceRequestOfficeDetailResourceCollection;
 use App\Repositories\Contracts\ServiceRequestOfficeDetailRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class ServiceRequestOfficeDetailController extends Controller
 {
@@ -31,9 +32,12 @@ class ServiceRequestOfficeDetailController extends Controller
      *
      * @param IndexRequest $request
      * @return ServiceRequestOfficeDetailResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [ServiceRequestOfficeDetail::class, $request->get('propertyId')]);
+
         $serviceRequestOfficeDetails = $this->serviceRequestOfficeDetailRepository->findBy($request->all());
 
         return new ServiceRequestOfficeDetailResourceCollection($serviceRequestOfficeDetails);
@@ -44,9 +48,12 @@ class ServiceRequestOfficeDetailController extends Controller
      *
      * @param  StoreRequest  $request
      * @return ServiceRequestOfficeDetailResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [ServiceRequestOfficeDetail::class, $request->get('propertyId')]);
+
         $serviceRequestOfficeDetail = $this->serviceRequestOfficeDetailRepository->setServiceRequestOfficeDetail($request->all());
 
         return new ServiceRequestOfficeDetailResource($serviceRequestOfficeDetail);
@@ -57,9 +64,12 @@ class ServiceRequestOfficeDetailController extends Controller
      *
      * @param ServiceRequestOfficeDetail $serviceRequestOfficeDetail
      * @return ServiceRequestOfficeDetailResource
+     * @throws AuthorizationException
      */
     public function show(ServiceRequestOfficeDetail $serviceRequestOfficeDetail)
     {
+        $this->authorize('show', $serviceRequestOfficeDetail);
+
         return new ServiceRequestOfficeDetailResource($serviceRequestOfficeDetail);
     }
 
@@ -69,9 +79,12 @@ class ServiceRequestOfficeDetailController extends Controller
      * @param UpdateRequest $request
      * @param ServiceRequestOfficeDetail $serviceRequestOfficeDetail
      * @return ServiceRequestOfficeDetailResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ServiceRequestOfficeDetail $serviceRequestOfficeDetail)
     {
+        $this->authorize('update', $serviceRequestOfficeDetail);
+
         $serviceRequestOfficeDetail = $this->serviceRequestOfficeDetailRepository->update($serviceRequestOfficeDetail,$request->all());
 
         return new ServiceRequestOfficeDetailResource($serviceRequestOfficeDetail);
@@ -82,9 +95,12 @@ class ServiceRequestOfficeDetailController extends Controller
      *
      * @param ServiceRequestOfficeDetail $serviceRequestOfficeDetail
      * @return \Illuminate\Http\Response
+     * @throws AuthorizationException
      */
     public function destroy(ServiceRequestOfficeDetail $serviceRequestOfficeDetail)
     {
+        $this->authorize('update', $serviceRequestOfficeDetail);
+
         $this->serviceRequestOfficeDetailRepository->delete($serviceRequestOfficeDetail);
 
         return response()->json(null,204);
